pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-bot"
version = "0.1.0"
description = "simple music bot"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "discord.py>=2.3.0",
    "PyNaCl>=1.5.0",
    "davey>=0.0.0",
    "yt-dlp>=2026.03.03",
    "sqlalchemy[asyncio]>=2.0.0",
    "psycopg[binary]>=3.3.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "types-yt-dlp",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
asyncio_mode = "auto"
markers = [
  "unit: fast isolated tests",
  "integration: tests that touch external systems (db, network, discord)",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]
extend-exclude = [".venv", "dist", "build", "__pycache__", ".mypy_cache", ".ruff_cache"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = ["E501", "E203"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.mypy]
python_version = "3.13"
mypy_path = ["src"]
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
no_implicit_optional = true
